+++ script.js
@@ -1,44 +1,30 @@
-(() => {
-  const btn = document.querySelector('[data-form-btn]');
+import checkComplete from './components/checkComplete.js';
+import deleteIcon from './components/deleteIcon.js';
 
-  const createTask = (evento) => {
-    evento.preventDefault();
-    const input = document.querySelector('[data-form-input]');
-    const value = input.value;
-    const list = document.querySelector('[data-list]');
-    const task = document.createElement('li');
-    task.classList.add('card');
-    input.value = '';
-    //backticks
-    const taskContent = document.createElement('div');
+const btn = document.querySelector('[data-form-btn]');
 
-    const titleTask = document.createElement('span');
-    titleTask.classList.add('task');
-    titleTask.innerText = value;
-    taskContent.appendChild(checkComplete());
-    taskContent.appendChild(titleTask);
+const createTask = (evento) => {
+  evento.preventDefault();
+  const input = document.querySelector('[data-form-input]');
+  const value = input.value;
+  const list = document.querySelector('[data-list]');
+  const task = document.createElement('li');
+  task.classList.add('card');
+  input.value = '';
+  //backticks
+  const taskContent = document.createElement('div');
 
-    const content = `
-    <i class="fas fa-trash-alt trashIcon icon"></i>`;
-    // task.innerHTML = content;
-    task.appendChild(taskContent);
-    list.appendChild(task);
-  };
+  const titleTask = document.createElement('span');
+  titleTask.classList.add('task');
+  titleTask.innerText = value;
+  taskContent.appendChild(checkComplete());
+  taskContent.appendChild(titleTask);
+  // task.innerHTML = content;
 
-  //Arrow functions o funciones anonimas
-  btn.addEventListener('click', createTask);
+  task.appendChild(taskContent);
+  task.appendChild(deleteIcon());
+  list.appendChild(task);
+};
 
-  const checkComplete = () => {
-    const i = document.createElement('i');
-    i.classList.add('far', 'fa-check-square', 'icon');
-    i.addEventListener('click', completeTask);
-    return i;
-  };
-  // Immediately invoked function expression IIFE
-  const completeTask = (event) => {
-    const element = event.target;
-    element.classList.toggle('fas');
-    element.classList.toggle('completeIcon');
-    element.classList.toggle('far');
-  };
-})();
+//Arrow functions o funciones anonimas
+btn.addEventListener('click', createTask);
